# Conflicts:
#	src/main/java/app/Controller/IngredientController.java
#	src/main/java/app/Controller/SousController.java
#	src/main/java/app/Controller/UserController.java
#	src/main/java/app/entity/InCard.java
#	src/main/java/app/entity/Ingredient.java
#	src/main/java/app/entity/Pizza.java
#	src/main/java/app/entity/PizzaOrder.java
#	src/main/java/app/entity/Sous.java
#	src/main/java/app/entity/User.java
#	src/main/resources/application.properties
#	target/classes/app/Controller/IngredientController.class
#	target/classes/app/Controller/MainPageController.class
#	target/classes/app/Controller/OrderController.class
#	target/classes/app/Controller/PizzaController.class
#	target/classes/app/Controller/UserController.class
#	target/classes/app/entity/Ingredient.class
#	target/classes/app/repo/OrderRepo.class

+++ src/main/java/app/repo/IngredientRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.Ingredient;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IngredientRepo extends CrudRepository<Ingredient, Long> {
+public interface IngredientRepo extends CrudRepository<Ingredient, Integer> {
 }
