+++ src/main/java/app/repo/PizzaRepo.java
@@ -3,5 +3,5 @@ package app.repo;
 import app.entity.Pizza;
 import org.springframework.data.repository.CrudRepository;
 
-public interface PizzaRepo extends CrudRepository<Pizza, Long> {
+public interface PizzaRepo extends CrudRepository<Pizza, Integer> {
 }
