+++ src/main/java/app/service/InCardService.java
@@ -1,9 +1,7 @@
 package app.service;
 
 import app.entity.InCard;
-import app.entity.InCard;
 import app.repo.InCardRepo;
-import app.repo.IngredientRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -21,15 +19,13 @@ public class InCardService {
     }
 
     private void deleteInCardByID(int id) {
-        inCardRepo.deleteById((long) id);
+        inCardRepo.deleteById(id);
     }
 
-    private void deleteInCard(InCard inCard) {
-        inCardRepo.delete(inCard);
-    }
+    private void deleteInCard(InCard inCard) { inCardRepo.delete(inCard); }
 
     private InCard getInCard(int id) {
-        Optional<InCard> optInCard = inCardRepo.findById((long) id);
+        Optional<InCard> optInCard = inCardRepo.findById(id);
         return optInCard.orElseGet(InCard::new);
     }
